update js cookies
