+++ package.json
@@ -14,6 +14,7 @@
     "chart.js": "^4.4.4",
     "dotenv": "^16.4.5",
     "formik": "^2.4.6",
+    "js-cookie": "^3.0.5",
     "moment": "^2.30.1",
     "next": "14.2.5",
     "react": "^18",
@@ -24,6 +25,7 @@
     "yup": "^1.4.0"
   },
   "devDependencies": {
+    "@types/js-cookie": "^3.0.6",
     "@types/next-auth": "^3.15.0",
     "@types/node": "^20",
     "@types/react": "^18",
