+++ src/hooks/useUserCookies.ts
@@ -1,54 +1,51 @@
-import { useCookies } from "react-cookie";
+import Cookies from "js-cookie";
 import { IUserGooglePut, IUserLogin } from "@/interfaces/RequestInterfaces";
 
 const googleUserKey = "googleUser";
 const regularUserKey = "regularUser";
 
 export function useUserCookies() {
-  const [cookies, setCookie, removeCookie] = useCookies([
-    googleUserKey,
-    regularUserKey,
-  ]);
-
+  // Guardar el usuario de Google en la cookie
   const saveGoogleUser = (newGoogleUser: IUserGooglePut) => {
     const dataToString = JSON.stringify(newGoogleUser);
-    setCookie(googleUserKey, dataToString, {
-      path: "/",
-      maxAge: 60 * 60 * 24 * 7, // 7 días
-    });
+    Cookies.set(googleUserKey, dataToString, { expires: 7 }); // expira en 7 días
   };
 
+  // Guardar un usuario regular en la cookie
   const saveRegularUser = (newRegularUser: IUserLogin) => {
     const dataToString = JSON.stringify(newRegularUser);
-    setCookie(regularUserKey, dataToString, {
-      path: "/",
-      maxAge: 60 * 60 * 24 * 7, // 7 días
-    });
+    Cookies.set(regularUserKey, dataToString, { expires: 7 }); // expira en 7 días
   };
 
-  const getRegularUser = (): IUserLogin | null => {
-    const getData = cookies[regularUserKey];
-    if (getData) return JSON.parse(getData);
-    else {
-      console.error("Regular User not found in cookies");
+  // Obtener el usuario de Google desde la cookie
+  const getGoogleUser = (): IUserGooglePut | null => {
+    const googleUser = Cookies.get(googleUserKey);
+    if (googleUser) {
+      try {
+        return JSON.parse(googleUser);
+      } catch (error) {
+        console.error("Error parsing Google User from cookies:", error);
+        return null;
+      }
+    } else {
       return null;
     }
   };
 
+  // Borrar el usuario de Google de la cookie
   const deleteGoogleUser = () => {
-    removeCookie(googleUserKey, { path: "/" });
+    Cookies.remove(googleUserKey);
   };
 
+  // Borrar el usuario regular de la cookie
   const deleteRegularUser = () => {
-    removeCookie(regularUserKey, { path: "/" });
+    Cookies.remove(regularUserKey);
   };
 
   return {
     saveGoogleUser,
     saveRegularUser,
-
-    getRegularUser,
-    deleteGoogleUser,
+    getGoogleUser,
     deleteRegularUser,
   };
 }
